Code cleanup - Inline CommandLineOpts into SpeedTestCmdLineOpts.
